refactor: encapsulate transaction processing with a typestate-based processor
