+++ stacks-codec/src/transaction.rs
@@ -2535,6 +2535,20 @@ pub struct StacksMicroblockHeader {
     pub signature: MessageSignature,
 }
 
+/// Signer relationship recovered from two valid microblock-header signatures.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum MicroblockSignerMatch {
+    /// Both headers were signed by the same key.
+    Common(Hash160),
+    /// The headers were signed by different keys.
+    Different {
+        /// Signer recovered from the first header.
+        first: Hash160,
+        /// Signer recovered from the second header.
+        second: Hash160,
+    },
+}
+
 impl StacksMessageCodec for StacksMicroblockHeader {
     fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
         self.serialize(fd, false)
@@ -2613,6 +2627,18 @@ impl StacksMicroblockHeader {
 
         pubk.set_compressed(true);
         Ok(Hash160::from_node_public_key(&pubk))
+    }
+
+    /// Recovers and compares the signers of two microblock headers.
+    pub fn recover_signer_match(&self, other: &Self) -> Result<MicroblockSignerMatch, AuthError> {
+        let first = self.check_recover_pubkey()?;
+        let second = other.check_recover_pubkey()?;
+
+        Ok(if first == second {
+            MicroblockSignerMatch::Common(first)
+        } else {
+            MicroblockSignerMatch::Different { first, second }
+        })
     }
 
     pub fn verify(&self, pubk_hash: &Hash160) -> Result<(), AuthError> {
@@ -4429,6 +4455,41 @@ mod tests {
         header.consensus_serialize(&mut buf).unwrap();
         let decoded = StacksMicroblockHeader::consensus_deserialize(&mut &buf[..]).unwrap();
         assert_eq!(decoded, header);
+    }
+
+    #[test]
+    fn microblock_headers_recover_signer_match() {
+        let signer = StacksPrivateKey::random();
+        let other_signer = StacksPrivateKey::random();
+        let parent = BlockHeaderHash([0x77; 32]);
+
+        let mut first =
+            StacksMicroblockHeader::first_unsigned(&parent, &Sha512Trunc256Sum([0x11; 32]));
+        first.sign(&signer).unwrap();
+
+        let mut same_signer =
+            StacksMicroblockHeader::first_unsigned(&parent, &Sha512Trunc256Sum([0x22; 32]));
+        same_signer.sign(&signer).unwrap();
+
+        let mut different_signer =
+            StacksMicroblockHeader::first_unsigned(&parent, &Sha512Trunc256Sum([0x33; 32]));
+        different_signer.sign(&other_signer).unwrap();
+
+        let first_signer = first.check_recover_pubkey().unwrap();
+        assert_eq!(
+            first.recover_signer_match(&same_signer).unwrap(),
+            MicroblockSignerMatch::Common(first_signer.clone()),
+        );
+        assert_eq!(
+            first.recover_signer_match(&different_signer).unwrap(),
+            MicroblockSignerMatch::Different {
+                first: first_signer,
+                second: different_signer.check_recover_pubkey().unwrap(),
+            },
+        );
+
+        let unsigned = StacksMicroblockHeader::first_empty_unsigned(&parent);
+        assert!(first.recover_signer_match(&unsigned).is_err());
     }
 
     /// Every `TransactionAuthFlags` discriminant must serialize to a single
