+++ stackslib/src/chainstate/stacks/db/transactions/processing/tests/mod.rs
@@ -1,0 +1,20 @@
+// Copyright (C) 2026 Stacks Open Internet Foundation
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#[cfg(doctest)]
+mod typestate_compile_fail;
+
+// TODO: Move "e2e-style" transaction processing tests here which don't test other specific units
+// in the `transactions/processing` module.
