+++ stackslib/src/chainstate/stacks/miner.rs
@@ -44,6 +44,7 @@ use crate::chainstate::stacks::address::StacksAddressExtensions;
 use crate::chainstate::stacks::db::blocks::SetupBlockResult;
 use crate::chainstate::stacks::db::transactions::{
     finalize_failed_transaction, handle_clarity_runtime_error, ClarityRuntimeTxError,
+    TransactionProcessor,
 };
 use crate::chainstate::stacks::db::unconfirmed::UnconfirmedState;
 use crate::chainstate::stacks::db::{ChainstateTx, ClarityTx, StacksChainState};
@@ -770,6 +771,7 @@ impl TransactionResult {
 /// This is a defense-in-depth measure -- if these budgets are exceeded, that
 /// probably means there's an underlying bug in the VM or analysis engine that
 /// should be fixed.
+#[derive(Debug, Clone, Copy)]
 pub struct TransactionResourceBudgets {
     /// The budget that applies during clarity evalution, used both during
     /// contract deploy and contract call transactions.
@@ -781,15 +783,12 @@ pub struct TransactionResourceBudgets {
 }
 
 impl TransactionResourceBudgets {
-    pub fn new() -> Self {
+    /// Creates resource budgets with no configured limits.
+    pub fn unlimited() -> Self {
         Self {
             execution_budget: ResourceBudget::unlimited(),
             analysis_budget: ResourceBudget::unlimited(),
         }
-    }
-
-    pub fn unlimited() -> Self {
-        Self::new()
     }
 
     pub fn from_settings(settings: &BlockBuilderSettings) -> Self {
@@ -800,10 +799,10 @@ impl TransactionResourceBudgets {
         };
 
         Self {
-            execution_budget: ResourceBudget::new()
+            execution_budget: ResourceBudget::unlimited()
                 .with_max_duration(settings.max_execution_time)
                 .with_max_memory_use(memory_limit),
-            analysis_budget: ResourceBudget::new()
+            analysis_budget: ResourceBudget::unlimited()
                 .with_max_duration(settings.max_analysis_time)
                 .with_max_memory_use(memory_limit),
         }
@@ -1113,7 +1112,7 @@ impl<'a> StacksMicroblockBuilder<'a> {
     /// Returns Ok(TransactionResult::Problematic) if the transaction should be dropped from the mempool.
     /// Returns Err(e) if an error occurs during the function.
     ///
-    /// This calls `StacksChainState::process_transaction` and also checks certain pre-conditions
+    /// This configures a `TransactionProcessor` and also checks certain pre-conditions
     /// and handles errors.
     ///
     /// # Pre-Checks
@@ -1195,7 +1194,12 @@ impl<'a> StacksMicroblockBuilder<'a> {
 
         let quiet = !cfg!(test);
         let cost_before = clarity_tx.cost_so_far();
-        match StacksChainState::process_transaction(clarity_tx, &tx, quiet, None) {
+        match TransactionProcessor::from(&tx)
+            .using_clarity_tx(clarity_tx)
+            .with_unlimited_resource_policy()
+            .quiet(quiet)
+            .process()
+        {
             Ok((_fee, receipt)) => TransactionResult::success(&tx, receipt),
             Err(e) => finalize_failed_transaction(clarity_tx, &tx, &cost_before, e),
         }
@@ -1756,7 +1760,12 @@ impl StacksBlockBuilder {
         let quiet = !cfg!(test);
         if !self.anchored_done {
             // save
-            match StacksChainState::process_transaction(clarity_tx, tx, quiet, None) {
+            match TransactionProcessor::from(tx)
+                .using_clarity_tx(clarity_tx)
+                .with_unlimited_resource_policy()
+                .quiet(quiet)
+                .process()
+            {
                 Ok((fee, receipt)) => {
                     self.total_anchored_fees += fee;
                 }
@@ -1767,7 +1776,12 @@ impl StacksBlockBuilder {
 
             self.txs.push(tx.clone());
         } else {
-            match StacksChainState::process_transaction(clarity_tx, tx, quiet, None) {
+            match TransactionProcessor::from(tx)
+                .using_clarity_tx(clarity_tx)
+                .with_unlimited_resource_policy()
+                .quiet(quiet)
+                .process()
+            {
                 Ok((fee, receipt)) => {
                     self.total_streamed_fees += fee;
                 }
@@ -2602,13 +2616,17 @@ impl BlockBuilder for StacksBlockBuilder {
                 return TransactionResult::problematic(tx, Error::NetError(e));
             }
             let cost_before = clarity_tx.cost_so_far();
-            let (fee, receipt) =
-                match StacksChainState::process_transaction(clarity_tx, tx, quiet, None) {
-                    Ok((fee, receipt)) => (fee, receipt),
-                    Err(e) => {
-                        return finalize_failed_transaction(clarity_tx, tx, &cost_before, e);
-                    }
-                };
+            let (fee, receipt) = match TransactionProcessor::from(tx)
+                .using_clarity_tx(clarity_tx)
+                .with_unlimited_resource_policy()
+                .quiet(quiet)
+                .process()
+            {
+                Ok((fee, receipt)) => (fee, receipt),
+                Err(e) => {
+                    return finalize_failed_transaction(clarity_tx, tx, &cost_before, e);
+                }
+            };
             info!("Include tx";
                   "tx" => %tx.txid(),
                   "payload" => tx.payload.name(),
@@ -2646,13 +2664,17 @@ impl BlockBuilder for StacksBlockBuilder {
                 return TransactionResult::problematic(tx, Error::NetError(e));
             }
             let cost_before = clarity_tx.cost_so_far();
-            let (fee, receipt) =
-                match StacksChainState::process_transaction(clarity_tx, tx, quiet, None) {
-                    Ok((fee, receipt)) => (fee, receipt),
-                    Err(e) => {
-                        return finalize_failed_transaction(clarity_tx, tx, &cost_before, e);
-                    }
-                };
+            let (fee, receipt) = match TransactionProcessor::from(tx)
+                .using_clarity_tx(clarity_tx)
+                .with_unlimited_resource_policy()
+                .quiet(quiet)
+                .process()
+            {
+                Ok((fee, receipt)) => (fee, receipt),
+                Err(e) => {
+                    return finalize_failed_transaction(clarity_tx, tx, &cost_before, e);
+                }
+            };
             debug!(
                 "Include tx {} ({}) in microblock",
                 tx.txid(),
