+++ stackslib/src/chainstate/stacks/mod.rs
@@ -415,16 +415,17 @@ impl Error {
 }
 
 pub use stacks_codec::transaction::{
-    AssetInfo, AssetInfoID, AuthError, CoinbasePayload, FungibleConditionCode, MultisigHashMode,
-    MultisigSpendingCondition, NonfungibleConditionCode, OrderIndependentMultisigHashMode,
-    OrderIndependentMultisigSpendingCondition, PostConditionPrincipal, PostConditionPrincipalID,
-    PoxConditionCode, SinglesigHashMode, SinglesigSpendingCondition, StacksMicroblockHeader,
-    StacksTransaction, TenureChangeCause, TenureChangeError, TenureChangePayload,
-    TokenTransferMemo, TransactionAnchorMode, TransactionAuth, TransactionAuthField,
-    TransactionAuthFieldID, TransactionAuthFlags, TransactionAuthVerificationMode,
-    TransactionContractCall, TransactionPayload, TransactionPayloadID, TransactionPostCondition,
-    TransactionPostConditionMode, TransactionPublicKeyEncoding, TransactionSmartContract,
-    TransactionSpendingCondition, TransactionVersion,
+    AssetInfo, AssetInfoID, AuthError, CoinbasePayload, FungibleConditionCode,
+    MicroblockSignerMatch, MultisigHashMode, MultisigSpendingCondition, NonfungibleConditionCode,
+    OrderIndependentMultisigHashMode, OrderIndependentMultisigSpendingCondition,
+    PostConditionPrincipal, PostConditionPrincipalID, PoxConditionCode, SinglesigHashMode,
+    SinglesigSpendingCondition, StacksMicroblockHeader, StacksTransaction, TenureChangeCause,
+    TenureChangeError, TenureChangePayload, TokenTransferMemo, TransactionAnchorMode,
+    TransactionAuth, TransactionAuthField, TransactionAuthFieldID, TransactionAuthFlags,
+    TransactionAuthVerificationMode, TransactionContractCall, TransactionPayload,
+    TransactionPayloadID, TransactionPostCondition, TransactionPostConditionMode,
+    TransactionPublicKeyEncoding, TransactionSmartContract, TransactionSpendingCondition,
+    TransactionVersion,
 };
 
 #[derive(Debug, Clone, PartialEq)]
