+++ stackslib/src/clarity_vm/clarity.rs
@@ -29,6 +29,7 @@ use clarity::vm::database::{
 };
 use clarity::vm::errors::VmExecutionError;
 use clarity::vm::events::{STXEventType, STXMintEventData};
+use clarity::vm::hooks::EvalHook;
 use clarity::vm::representations::SymbolicExpression;
 use clarity::vm::resource_limiter::ResourceBudget;
 use clarity::vm::types::{PrincipalData, QualifiedContractIdentifier, Value};
@@ -47,10 +48,10 @@ use crate::chainstate::stacks::boot::{
     SIGNERS_BODY, SIGNERS_DB_0_BODY, SIGNERS_DB_1_BODY, SIGNERS_NAME, SIGNERS_VOTING_BODY,
     SIGNERS_VOTING_NAME, SIP_031_NAME,
 };
+use crate::chainstate::stacks::db::transactions::{self, TransactionProcessor};
 use crate::chainstate::stacks::db::{StacksAccount, StacksChainState};
 use crate::chainstate::stacks::events::{StacksTransactionEvent, StacksTransactionReceipt};
 use crate::chainstate::stacks::index::marf::MARF;
-use crate::chainstate::stacks::miner::TransactionResourceBudgets;
 use crate::chainstate::stacks::{
     Error as ChainstateError, StacksMicroblockHeader, StacksTransaction, TransactionPayload,
     TransactionSmartContract, TransactionVersion,
@@ -62,6 +63,17 @@ use crate::core::{StacksEpoch, StacksEpochId, FIRST_STACKS_BLOCK_ID, GENESIS_EPO
 use crate::util_lib::boot::{boot_code_acc, boot_code_addr, boot_code_id, boot_code_tx_auth};
 use crate::util_lib::db::Error as DatabaseError;
 use crate::util_lib::strings::StacksString;
+
+fn process_boot_transaction_payload(
+    clarity_tx: &mut ClarityTransactionConnection,
+    tx: &StacksTransaction,
+    origin_account: &StacksAccount,
+) -> Result<StacksTransactionReceipt, ChainstateError> {
+    TransactionProcessor::from(tx)
+        .using_clarity_transaction(clarity_tx, origin_account)
+        .with_unlimited_resource_policy()
+        .process_payload()
+}
 
 pub const SIP_031_INITIAL_MINT: u128 = 200_000_000_000_000;
 
@@ -139,6 +151,8 @@ pub struct ClarityTransactionConnection<'a, 'b> {
     /// Per-tx cache container which is attached to [`ClarityDatabase`] instances handed out by this
     /// connection.
     cache: ClarityExecutionCache,
+    /// Optional evaluation hook scoped to this transaction.
+    eval_hook: Option<&'b mut dyn EvalHook>,
 }
 
 /// Unified API common to all MARF stores
@@ -277,7 +291,13 @@ impl<'a, 'b> ClarityTransactionConnection<'a, 'b> {
             chain_id,
             epoch,
             cache: ClarityExecutionCache::default(),
+            eval_hook: None,
         }
+    }
+
+    /// Attaches a Clarity evaluation hook to this transaction connection.
+    pub fn set_eval_hook(&mut self, eval_hook: &'b mut dyn EvalHook) {
+        self.eval_hook = Some(eval_hook);
     }
 }
 
@@ -1162,13 +1182,8 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
         let receipt = self.as_free_transaction(|tx_conn| {
             info!("Instantiate {} contract", &contract_id);
-            StacksChainState::process_transaction_payload(
-                tx_conn,
-                &contract_tx,
-                &boot_code_account,
-                &TransactionResourceBudgets::unlimited(),
-            )
-            .expect("FATAL: Failed to process boot contract initialization")
+            process_boot_transaction_payload(tx_conn, &contract_tx, &boot_code_account)
+                .expect("FATAL: Failed to process boot contract initialization")
         });
 
         if receipt.result != Value::okay_true() || receipt.post_condition_aborted {
@@ -1236,13 +1251,8 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
                 // initialize with a synthetic transaction
                 debug!("Instantiate .costs-2 contract");
-                StacksChainState::process_transaction_payload(
-                    tx_conn,
-                    &costs_2_contract_tx,
-                    &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
-                )
-                .expect("FATAL: Failed to process PoX 2 contract initialization")
+                process_boot_transaction_payload(tx_conn, &costs_2_contract_tx, &boot_code_account)
+                    .expect("FATAL: Failed to process PoX 2 contract initialization")
             });
 
             if initialization_receipt.result != Value::okay_true()
@@ -1346,11 +1356,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
                 // initialize with a synthetic transaction
                 debug!("Instantiate {} contract", &pox_2_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &pox_2_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process PoX 2 contract initialization");
 
@@ -1417,11 +1426,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
                 // initialize with a synthetic transaction
                 debug!("Instantiate .costs-3 contract");
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &costs_3_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process costs-3 contract initialization");
 
@@ -1585,11 +1593,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
             let pox_3_initialization_receipt = self.as_transaction(|tx_conn| {
                 // initialize with a synthetic transaction
                 debug!("Instantiate {} contract", &pox_3_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &pox_3_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process PoX 3 contract initialization");
 
@@ -1703,11 +1710,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
             let pox_4_initialization_receipt = self.as_transaction(|tx_conn| {
                 // initialize with a synthetic transaction
                 debug!("Instantiate {} contract", &pox_4_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &pox_4_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process PoX 4 contract initialization");
 
@@ -1762,11 +1768,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
             let signers_initialization_receipt = self.as_transaction(|tx_conn| {
                 // initialize with a synthetic transaction
                 debug!("Instantiate {} contract", &signers_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &signers_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process .signers contract initialization");
                 receipt
@@ -1808,11 +1813,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
                     let signers_db_receipt = self.as_transaction(|tx_conn| {
                         // initialize with a synthetic transaction
                         debug!("Instantiate .{} contract", &signers_name);
-                        let receipt = StacksChainState::process_transaction_payload(
+                        let receipt = process_boot_transaction_payload(
                             tx_conn,
                             &signers_contract_tx,
                             &boot_code_account,
-                            &TransactionResourceBudgets::unlimited(),
                         )
                         .expect("FATAL: Failed to process .signers DB contract initialization");
                         receipt
@@ -1847,11 +1851,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
             let signers_voting_initialization_receipt = self.as_transaction(|tx_conn| {
                 // initialize with a synthetic transaction
                 debug!("Instantiate {} contract", &signers_voting_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &signers_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process .signers-voting contract initialization");
                 receipt
@@ -1978,11 +1981,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
             let mut sip_031_initialization_receipt = self.as_transaction(|tx_conn| {
                 // initialize with a synthetic transaction
                 info!("Instantiate {} contract", &sip_031_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &sip_031_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process .sip-031 contract initialization");
                 receipt
@@ -2092,11 +2094,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
                 // initialize with a synthetic transaction
                 info!("Instantiate .costs-4 contract");
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &costs_4_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process costs-4 contract initialization");
 
@@ -2190,11 +2191,10 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
 
             let pox_5_initialization_receipt = self.as_transaction(|tx_conn| {
                 debug!("Instantiate {} contract", &pox_5_contract_id);
-                let receipt = StacksChainState::process_transaction_payload(
+                let receipt = process_boot_transaction_payload(
                     tx_conn,
                     &pox_5_contract_tx,
                     &boot_code_account,
-                    &TransactionResourceBudgets::unlimited(),
                 )
                 .expect("FATAL: Failed to process .pox-5 contract initialization");
 
@@ -2420,6 +2420,10 @@ impl TransactionConnection for ClarityTransactionConnection<'_, '_> {
                     self.epoch,
                 );
 
+                if let Some(eval_hook) = self.eval_hook.as_deref_mut() {
+                    vm_env.add_eval_hook(eval_hook);
+                }
+
                 let result = to_do(&mut vm_env);
                 let (mut db, cost_track) = vm_env
                     .destruct()
@@ -2520,7 +2524,7 @@ impl ClarityTransactionConnection<'_, '_> {
                 vm_env
                     .execute_in_env(sender.clone(), None, None, |exec_state, invoke_ctx| {
                         exec_state.run_as_transaction(invoke_ctx, |exec_state, invoke_ctx| {
-                            StacksChainState::handle_poison_microblock(
+                            transactions::processing::poison_microblock::handle(
                                 exec_state,
                                 invoke_ctx,
                                 mblock_header_1,
@@ -2631,6 +2635,7 @@ mod tests {
     use clarity::types::chainstate::{BurnchainHeaderHash, SortitionId, StacksAddress};
     use clarity::vm::analysis::errors::RuntimeCheckErrorKind;
     use clarity::vm::database::{ClarityBackingStore, STXBalance, SqliteConnection};
+    use clarity::vm::hooks::trace::CallTraceHook;
     use clarity::vm::test_util::{TEST_BURN_STATE_DB, TEST_HEADER_DB};
     use clarity::vm::types::{StandardPrincipalData, TupleData, Value};
     use clarity::vm::ClarityName;
@@ -3453,33 +3458,15 @@ mod tests {
             );
 
             conn.as_transaction(|clarity_tx| {
-                let receipt = StacksChainState::process_transaction_payload(
-                    clarity_tx,
-                    &tx1,
-                    &account,
-                    &TransactionResourceBudgets::unlimited(),
-                )
-                .unwrap();
+                let receipt = process_boot_transaction_payload(clarity_tx, &tx1, &account).unwrap();
                 assert!(receipt.post_condition_aborted);
             });
             conn.as_transaction(|clarity_tx| {
-                StacksChainState::process_transaction_payload(
-                    clarity_tx,
-                    &tx2,
-                    &account,
-                    &TransactionResourceBudgets::unlimited(),
-                )
-                .unwrap();
+                process_boot_transaction_payload(clarity_tx, &tx2, &account).unwrap();
             });
 
             conn.as_transaction(|clarity_tx| {
-                let receipt = StacksChainState::process_transaction_payload(
-                    clarity_tx,
-                    &tx3,
-                    &account,
-                    &TransactionResourceBudgets::unlimited(),
-                )
-                .unwrap();
+                let receipt = process_boot_transaction_payload(clarity_tx, &tx3, &account).unwrap();
 
                 assert!(receipt.post_condition_aborted);
             });
@@ -3735,8 +3722,12 @@ mod tests {
                 .unwrap();
         });
 
-        // tx2: first call misses, second call hits.
-        conn.as_transaction(|tx| {
+        // tx2: first call misses, second call hits. The same eval hook must be
+        // reborrowed into both environments created by these calls.
+        let mut call_trace = CallTraceHook::new();
+        {
+            let mut tx = conn.start_transaction_processing();
+            tx.set_eval_hook(&mut call_trace);
             assert_eq!(tx.cache.contracts.hits(), 0);
             assert_eq!(tx.cache.contracts.misses(), 0);
 
@@ -3769,7 +3760,13 @@ mod tests {
                 tx.cache.contracts.hits() > hits_after_first,
                 "second call should hit the cache",
             );
-        });
+            tx.commit().unwrap();
+        }
+        assert_eq!(
+            call_trace.calls().len(),
+            2,
+            "both sequential environments should use the attached eval hook",
+        );
 
         // tx3: fresh tx must see counters reset and still work.
         conn.as_transaction(|tx| {
