+++ stackslib/src/net/api/postblock_proposal.rs
@@ -746,14 +746,14 @@ impl NakamotoBlockProposal {
         } else {
             None
         };
-        let resource_budgets = TransactionResourceBudgets::new()
+        let resource_budgets = TransactionResourceBudgets::unlimited()
             .with_analysis_budget(
-                ResourceBudget::new()
+                ResourceBudget::unlimited()
                     .with_max_duration(Some(per_tx_max_analysis_time))
                     .with_max_memory_use(max_tx_mem_bytes_opt),
             )
             .with_execution_budget(
-                ResourceBudget::new()
+                ResourceBudget::unlimited()
                     .with_max_duration(Some(per_tx_max_execution_time))
                     .with_max_memory_use(max_tx_mem_bytes_opt),
             );
